Testing for VNH modules

+++ Software/AktuatorBoard/src/modules.h
@@ -8,7 +8,7 @@ void switch_relay(uint8_t module);
 int status_relay(uint8_t module);
 
 void setup_l293d();
-void write_l293d(uint8_t pwm, uint8_t dir);
+void write_l293d(uint8_t pwm, int dir);
 void stop_l293d(void);
 
 void setup_stepper(uint8_t rpm, uint8_t module);
